Reduce the cache miss rate
In order to improve the performance, begin with reducing the cache miss.

At first, for compressing the sizes of structure, indicate the last name
into a new structure, then create a hash table and put the new structure
inside, so sizes of each entry become to 8 byte from 136 byte, because
it only has a pointer inside.

Next, every structure aligned to cache line, so it would avoid a data
cross two cache blocks and each data will be access cache once.

Cache miss rate was less and the cost of find name was much faster after complete above modifying.